requirements.txt
# Podstawa ML i obsługa GPU
torch>=2.1.0
torchvision
torchaudio

# Modele Hugging Face
transformers>=4.40.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
sentencepiece
protobuf
einops

# Przetwarzanie danych i metryki
pandas
numpy
scikit-learn
tqdm
pillow

# Wizualizacja
matplotlib
seaborn

# Pobieranie danych
kaggle